move endpoint implementations in separate files and rename
