+++ internal/server/codeSystem.go
@@ -1,0 +1,31 @@
+package server
+
+import (
+	"context"
+	"miracummapper/internal/api"
+)
+
+// AddCodeSystem implements api.StrictServerInterface.
+func (s *Server) AddCodeSystem(ctx context.Context, request api.AddCodeSystemRequestObject) (api.AddCodeSystemResponseObject, error) {
+	panic("unimplemented")
+}
+
+// DeleteCodeSystem implements api.StrictServerInterface.
+func (s *Server) DeleteCodeSystem(ctx context.Context, request api.DeleteCodeSystemRequestObject) (api.DeleteCodeSystemResponseObject, error) {
+	panic("unimplemented")
+}
+
+// GetAllCodeSystems implements api.StrictServerInterface.
+func (s *Server) GetAllCodeSystems(ctx context.Context, request api.GetAllCodeSystemsRequestObject) (api.GetAllCodeSystemsResponseObject, error) {
+	panic("unimplemented")
+}
+
+// GetCodeSystem implements api.StrictServerInterface.
+func (s *Server) GetCodeSystem(ctx context.Context, request api.GetCodeSystemRequestObject) (api.GetCodeSystemResponseObject, error) {
+	panic("unimplemented")
+}
+
+// UpdateCodeSystem implements api.StrictServerInterface.
+func (s *Server) UpdateCodeSystem(ctx context.Context, request api.UpdateCodeSystemRequestObject) (api.UpdateCodeSystemResponseObject, error) {
+	panic("unimplemented")
+}
