+++ internal/server/mapping.go
@@ -1,0 +1,31 @@
+package server
+
+import (
+	"context"
+	"miracummapper/internal/api"
+)
+
+// AddMapping implements api.StrictServerInterface.
+func (s *Server) AddMapping(ctx context.Context, request api.AddMappingRequestObject) (api.AddMappingResponseObject, error) {
+	panic("unimplemented")
+}
+
+// DeleteMapping implements api.StrictServerInterface.
+func (s *Server) DeleteMapping(ctx context.Context, request api.DeleteMappingRequestObject) (api.DeleteMappingResponseObject, error) {
+	panic("unimplemented")
+}
+
+// GetAllMappings implements api.StrictServerInterface.
+func (s *Server) GetAllMappings(ctx context.Context, request api.GetAllMappingsRequestObject) (api.GetAllMappingsResponseObject, error) {
+	panic("unimplemented")
+}
+
+// GetMappingByID implements api.StrictServerInterface.
+func (s *Server) GetMappingByID(ctx context.Context, request api.GetMappingByIDRequestObject) (api.GetMappingByIDResponseObject, error) {
+	panic("unimplemented")
+}
+
+// UpdateMapping implements api.StrictServerInterface.
+func (s *Server) UpdateMapping(ctx context.Context, request api.UpdateMappingRequestObject) (api.UpdateMappingResponseObject, error) {
+	panic("unimplemented")
+}
